middleware kind of works, pre splitting functions inbetween middleware and user model
