+++ auth/auth-router.js
@@ -7,7 +7,8 @@ const Users = require("../users/users-model");
 const secrets = require("../api/secrets");
 const check = require('../middleware/index')
 
-router.post("/register", check.validateUser, (req, res) => {
+router.post("/register", check.validateUser, check.handleUsersLocations, (req, res) => {
+    console.log('does user location obj get passed', userLocation)
   let user = req.body; // username, password
 
   // rounds are 2 to the N times
